feat: 집단지성 수정 service 구현 #43

+++ src/main/java/com/hogu/am_i_hogu/domain/comment/service/CommentCreateService.java
@@ -140,7 +140,7 @@ public class CommentCreateService {
             Comment comment
     ) {
         boolean isPostWriter = writer.getId().equals(post.getWriter().getId());
-        int helpfulMarkCount = getHelpfulMarkCount(comment.getId());
+        long helpfulMarkCount = getHelpfulMarkCount(comment.getId());
         Long parentId = comment.getParentComment() == null
                 ? null
                 : comment.getParentComment().getId();
@@ -163,7 +163,7 @@ public class CommentCreateService {
         );
     }
 
-    private int getHelpfulMarkCount(Long commentId) {
-        return Math.toIntExact(commentHelpfulMarkRepository.countById_CommentId(commentId));
+    private long getHelpfulMarkCount(Long commentId) {
+        return commentHelpfulMarkRepository.countById_CommentId(commentId);
     }
 }
